`help` command in ZSH like in BASH
By default, ZSH does not provide `help` command. There is a `run-help`
but it is aliased to `man` so only external commands and not shell
built-ins.

To have the functionality of `help` as it is in BASH (show information
about built-in commands), run-help needs to be loaded (multiple ones)
and an alias created.

For info see https://wiki.archlinux.org/title/Zsh#Help_command